+++ package.json
@@ -4,10 +4,10 @@
   "description": "A small tool to find unused Sass variables",
   "author": "XhmikosR <xhmikosr@gmail.com>",
   "license": "MIT",
-  "main": "api.js",
+  "main": "index.js",
   "bin": {
-    "find-unused-sass-variables" : "./index.js",
-    "fusv" : "./index.js"
+    "find-unused-sass-variables" : "./cli.js",
+    "fusv" : "./cli.js"
   },
   "scripts": {
     "eslint": "eslint .",
@@ -29,8 +29,7 @@
     "eslint": "^4.12.0"
   },
   "files": [
-    "src/findUnusedVars.js",
-    "api.js",
+    "cli.js",
     "index.js"
   ],
   "engines": {
